+++ projects/Mallard/src/download-edition/prepare-and-download-edition.ts
@@ -3,7 +3,7 @@ import { fetchCacheClear } from '../helpers/fetch'
 import { clearOldIssues, prepFileSystem } from '../helpers/files'
 import { cleanPushTrackingByDays } from '../push-notifications/push-tracking'
 import { largeDeviceMemory } from 'src/hooks/use-config-provider'
-import { downloadTodaysEdition } from 'src/download/downloadTodaysEdition'
+import { downloadTodaysEdition } from 'src/download-edition/download-todays-edition'
 
 const prepareAndDownloadTodaysEdition = async (
     client: ApolloClient<object>,
